requirements.txt
certifi==2020.12.5
chardet==4.0.0
idna==2.10
Jinja2==2.11.3
MarkupSafe==1.1.1
python-dateutil==2.8.1
python-dotenv==0.15.0
requests==2.25.1
schedule==0.6.0
selenium==3.141.0
six==1.15.0
urllib3==1.26.3
